#29 moved to partial file to reduce confusion between synchronous and asynchronous code.
